refactor(artifacts): extract artifact polling boundary
